validation videos specified in yaml file

+++ validator.py
@@ -12,8 +12,8 @@ def validate(ground_truth, estimation):
     no_estimation = np.isnan(estimation)
     false_alarms = np.sum(np.logical_and(no_annotation, np.logical_not(no_estimation)))
     missed_detections = np.sum(np.logical_and(np.logical_not(no_annotation), no_estimation))
-    false_alarm_per_frame = false_alarms/diff.shape[1]
-    missed_detection_per_frame = missed_detections/diff.shape[1]
+    false_alarm_per_frame = float(false_alarms/diff.shape[1])
+    missed_detection_per_frame = float(missed_detections/diff.shape[1])
     return {'Error per correctly detected frame': error_per_ok_frame, 'False alarm per frame': false_alarm_per_frame,
             'Missed detection per frame': missed_detection_per_frame}
